# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Manga Scraper API server.",
        "title": "Manga Scraper API",
        "termsOfService": "https://manga-scraper.hostinger.fourleaves.studio/terms",
        "contact": {
            "name": "API Support",
            "url": "https://manga-scraper.hostinger.fourleaves.studio/support",
            "email": "admin@fourleaves.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/scrape-requests/chapters/detail": {
            "put": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Create request to scrape chapter detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape-requests"
                ],
                "summary": "Create request to scrape chapter detail",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1Binding.PutScrapeChapterDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/scrape-requests/series/detail": {
            "put": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Create request to scrape series detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape-requests"
                ],
                "summary": "Create request to scrape series detail",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1Binding.PutScrapeSeriesDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1Binding.PutScrapeChapterDetail": {
            "type": "object",
            "required": [
                "chapter",
                "provider",
                "series"
            ],
            "properties": {
                "chapter": {
                    "type": "string",
                    "example": "reincarnator-chapter-27"
                },
                "provider": {
                    "type": "string",
                    "example": "asura"
                },
                "series": {
                    "type": "string",
                    "example": "reincarnator"
                }
            }
        },
        "v1Binding.PutScrapeSeriesDetail": {
            "type": "object",
            "required": [
                "provider",
                "series"
            ],
            "properties": {
                "provider": {
                    "type": "string",
                    "example": "asura"
                },
                "series": {
                    "type": "string",
                    "example": "reincarnator"
                }
            }
        },
        "v1Response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "detail": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}